Move startup stuff out of vulcan-core
