+++ packages/lesswrong/client.js
@@ -1,6 +1,6 @@
 import './client/vulcan-lib/main';
 import './lib/vulcan-lib';
-import './client/vulcan-core/start';
+import './client/start';
 
 // Make sure to register settings before everything else
 import './client/publicSettings'
