--- packages/lesswrong/server/vulcan-core/start.ts
@@ -1,7 +1,0 @@
-import { onStartup } from '../../lib/executionEnvironment';
-import { startSyncedCron } from '../cronUtil';
-
-onStartup(function() {
-  startSyncedCron();
-});
-
